modified:   cmd/avito-test-task/main.go 	modified:   internal/pkg/controller/http/v1/account.go 	modified:   internal/pkg/controller/http/v1/balance.go 	modified:   internal/pkg/controller/http/v1/controllers.go 	modified:   internal/pkg/entity/entitys.go 	modified:   internal/pkg/usecase/balance.go 	modified:   internal/pkg/usecase/storage/account_postgres.go 	modified:   internal/pkg/usecase/storage/balance_postgres.go 	modified:   internal/pkg/usecase/storage/storage.go 	modified:   internal/pkg/usecase/usecase.go

+++ internal/pkg/entity/entitys.go
@@ -5,15 +5,15 @@ import "time"
 type (
 	User struct {
 		ID        uint64    `json:"id"`
-		Balance   uint64    `json:"balance"`
+		Balance   float64   `json:"balance"`
 		CreatedDT time.Time `json:"created_dt"`
 	}
 
 	Operation struct {
-		ID            uint64 `json:"id"`
-		OperationType string `json:"operation_type"`
-		Amount        int64  `amount:"amount"`
-		Currency      string `json:"currency"`
-		CreatedDT     int64  `json:"created_dt"`
+		ID            uint64  `json:"id"`
+		OperationType string  `json:"operation_type"`
+		Amount        float64 `amount:"amount"`
+		Currency      string  `json:"currency"`
+		CreatedDT     int64   `json:"created_dt"`
 	}
 )
